Move Rust wallet out of `src` directory
This is to avoid so many levels of nesting.
