+++ waves/src/index.tsx
@@ -18,8 +18,6 @@ ReactDOM.render(
     document.getElementById("root"),
 );
 
-import("./wallet/pkg").then(wallet => (window as any).window = wallet);
-
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
