+++ waves/wallet/.cargo/config.toml
