+++ waves/wallet/.gitignore
