+++ waves/wallet/Cargo.toml
@@ -14,7 +14,7 @@ default = [ "console_error_panic_hook" ]
 anyhow = "1"
 conquer-once = "0.3"
 console_error_panic_hook = { version = "0.1.6", optional = true }
-elements-fun = { path = "../../../elements-fun", features = [ "serde" ] }
+elements-fun = { path = "../../elements-fun", features = [ "serde" ] }
 futures = "0.3"
 hex = "0.4"
 itertools = "0.9.0"
