+++ waves/wallet/src/cache_storage.rs
