+++ waves/wallet/src/lib.rs
