+++ waves/wallet/src/macros.rs
