+++ waves/wallet/src/typed_js_future.rs
