+++ waves/wallet/src/utils.rs
