+++ waves/wallet/src/wallet.rs
